MNT #594 callbacks
