+++ docs/source/api/_callbacks.rst
@@ -1,6 +1,10 @@
-
+=========
 Callbacks
----------
+=========
 
-.. automodule:: apstools.callbacks
+.. automodule:: apstools.callbacks.doc_collector
     :members:
+
+.. automodule:: apstools.callbacks.snapshot_report
+    :members:
+
