+++ ideas/README.md
@@ -1,0 +1,4 @@
+# README
+
+This directory is not necessarily current.  It's just ideas from a previous
+period.
